perf(ui): split out gallery settings popover components
This was taking over 15ms (!) to render each time a setting changed, wtf
